+++ OpenBudgeteer.Blazor/Pages/Report.razor.cs
@@ -2,8 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using ChartJs.Blazor.ChartJS.BarChart;
+using ApexCharts;
 using Microsoft.AspNetCore.Components;
+using OpenBudgeteer.Blazor.Common;
 using OpenBudgeteer.Blazor.ViewModels;
 using OpenBudgeteer.Core.Data.Contracts.Services;
 
@@ -12,21 +13,54 @@ namespace OpenBudgeteer.Blazor.Pages;
 public partial class Report : ComponentBase
 {
     [Inject] private IServiceManager ServiceManager { get; set; } = null!;
-
-    private BlazorReportPageViewModel _dataContext = null!;
-    private List<Tuple<string, BarConfig>> _monthBucketExpensesConfigsLeft = null!;
-    private List<Tuple<string, BarConfig>> _monthBucketExpensesConfigsRight = null!;
+    
+    private ApexChart<ReportRecord> MonthBalanceChart;
+    private ApexChart<ReportRecord> BankBalanceChart;
+    private ApexChart<ReportRecord> MonthIncomeExpensesChart;
+    private ApexChart<ReportRecord> YearIncomeExpensesChart;
+    private List<ApexChart<ReportRecord>> MonthBucketExpensesCharts = new();
+    private ApexChart<ReportRecord> InjectMonthBucketExpensesChart
+    {
+        set => MonthBucketExpensesCharts.Add(value);
+    }
+    
+    private ApexChartOptions<ReportRecord> MonthIncomeExpensesChartOptions = new()
+    {
+        Legend = new() { Show = false }
+    };
+    private ApexChartOptions<ReportRecord> YearIncomeExpensesChartOptions = new()
+    {
+        Legend = new() { Show = false }
+    };
+    
+    private ApexReportViewModel _apexContext = null!;
+    private List<Tuple<string, List<ReportRecord>>> _monthBucketExpensesConfigsLeft = null!;
+    private List<Tuple<string, List<ReportRecord>>> _monthBucketExpensesConfigsRight = null!;
 
     protected override async Task OnInitializedAsync()
     {
-        _monthBucketExpensesConfigsLeft = new List<Tuple<string, BarConfig>>();
-        _monthBucketExpensesConfigsRight = new List<Tuple<string, BarConfig>>();
+        _monthBucketExpensesConfigsLeft = new List<Tuple<string, List<ReportRecord>>>();
+        _monthBucketExpensesConfigsRight = new List<Tuple<string, List<ReportRecord>>>();
+        MonthBucketExpensesCharts = new();
+    
+        _apexContext = new ApexReportViewModel(ServiceManager);
+        await _apexContext.LoadDataAsync();
+    
+        var halfIndex = _apexContext.MonthBucketExpenses.Count / 2;
+        _monthBucketExpensesConfigsLeft.AddRange(_apexContext.MonthBucketExpenses.GetRange(0,halfIndex));
+        _monthBucketExpensesConfigsRight.AddRange(_apexContext.MonthBucketExpenses.GetRange(halfIndex,_apexContext.MonthBucketExpenses.Count - halfIndex));
+        
+        StateHasChanged();
+        var tasks = new List<Task>()
+        {
+            MonthBalanceChart.UpdateSeriesAsync(true),
+            BankBalanceChart.UpdateSeriesAsync(true),
+            MonthIncomeExpensesChart.UpdateSeriesAsync(true),
+            YearIncomeExpensesChart.UpdateSeriesAsync(true)
+        };
+        tasks.AddRange(MonthBucketExpensesCharts
+            .Select(monthBucketExpensesChart => monthBucketExpensesChart.UpdateSeriesAsync()));
 
-        _dataContext = new BlazorReportPageViewModel(ServiceManager);
-        await _dataContext.LoadDataAsync();
-
-        var halfIndex = _dataContext.MonthBucketExpensesConfigs.Count / 2;
-        _monthBucketExpensesConfigsLeft.AddRange(_dataContext.MonthBucketExpensesConfigs.ToList().GetRange(0,halfIndex));
-        _monthBucketExpensesConfigsRight.AddRange(_dataContext.MonthBucketExpensesConfigs.ToList().GetRange(halfIndex,_dataContext.MonthBucketExpensesConfigs.Count - halfIndex));
+        await Task.WhenAll(tasks);
     }
 }